+++ Funcky/Functional/RetryPolicies/DoNotRetryPolicy.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Funcky
+{
+    public sealed class DoNotRetryPolicy : IRetryPolicy
+    {
+        public int MaxRetries => 0;
+
+        public TimeSpan Duration(int onRetryCount) => TimeSpan.Zero;
+    }
+}